changed backgrounds for pages/homepage to be different, added bio to about page
