+++ application/src/main/resources/static/js/react-client/src/components/footer/style.module.scss
@@ -1,0 +1,3 @@
+footer {
+    margin-top: auto;
+}
