+++ application/src/main/resources/static/js/react-client/src/pages/artgallery-screen/index.js
@@ -1,9 +1,21 @@
-import React, { useEffect, useState } from "react";
+import React, { createContext, useEffect, useState } from "react";
 
 import Axios from "axios";
+import ReactSwitch from "react-switch";
 import style from './style.module.scss';
 
+export const ThemeContext = createContext(null);
+
 const ArtgalleryScreen = () => {
+
+    // light/dark mode
+    const [ theme, setTheme ] = useState("dark");
+    const toggleTheme = () => {
+    setTheme((curr) => (curr === "light" ? "dark" : "light"));
+    };
+     // light/dark mode
+
+    
     const [art, setArt] = useState(null);
     const [artTwo, setArtTwo] = useState(null);
     const [loading, setLoading] = useState(true);
@@ -36,6 +48,18 @@ const ArtgalleryScreen = () => {
             return () => clearTimeout(timer);
             }, [artTwo]);
     return (
+
+         // light/dark mode
+        <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      <div className="Artgallery" id={theme} >   
+      <div className={theme=== 'dark' ? style.dark : style.light}>
+      <div className="switch">
+          {/* <label> {theme === "light" ? "Light Mode" : "Dark Mode"}</label> */}
+        <ReactSwitch onChange={toggleTheme} checked={theme === "dark"}/>
+        </div>
+    {/* light/dark mode */}
+
+        
         loading ? <h3>Creating art...</h3> :
         <div className={style.artContainer}>
             <h1>The Metropolitan Museum of Art</h1>
@@ -48,6 +72,11 @@ const ArtgalleryScreen = () => {
           <h3>{artTwo.artistDisplayName}</h3>
           <img src={artTwo.primaryImage} />
         </div>
+        
+        {/* light/dark mode */}
+        </div>
+        </div>
+        </ThemeContext.Provider>
     );
 };
 export default ArtgalleryScreen;
