+++ application/src/main/resources/static/js/react-client/src/pages/artgallery-screen/style.module.scss
@@ -14,3 +14,14 @@
         text-align: center;
     }
 }
+
+.light {
+    background-image: url(../../assets/background/pages-light-background.png) !important;
+    background-size: cover;
+   }
+  
+.dark {
+    background-image: url(../../assets/background/pages-dark-background.png) !important;
+    color: white;
+    background-size: cover;
+   }
