+++ application/src/main/resources/static/js/react-client/src/pages/contact-screen/style.module.scss
@@ -1,0 +1,10 @@
+.light {
+    background-image: url(../../assets/background/pages-light-background.png) !important;
+    background-size: cover;
+   }
+  
+.dark {
+    background-image: url(../../assets/background/pages-dark-background.png) !important;
+    color: white;
+    background-size: cover;
+   }
