+++ application/src/main/resources/static/js/react-client/src/pages/math-screen/index.js
@@ -1,9 +1,26 @@
-import React from "react";
+import React, { createContext, useState } from "react";
+
+import ReactSwitch from "react-switch";
+import style from './style.module.scss';
+
+export const ThemeContext = createContext(null);
 
 const MathScreen = () => {
 
+    const [ theme, setTheme ] = useState("dark");
+    const toggleTheme = () => {
+      setTheme((curr) => (curr === "light" ? "dark" : "light"));
+    };
+    
     return (
 
+        <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      <div className="Contact" id={theme} >   
+      <div className={theme=== 'dark' ? style.dark : style.light}>
+      <div className="switch">
+          {/* <label> {theme === "light" ? "Light Mode" : "Dark Mode"}</label> */}
+        <ReactSwitch onChange={toggleTheme} checked={theme === "dark"}/>
+        </div>
 
         <div>
             <h2>This is the math screen</h2>
@@ -34,6 +51,10 @@ const MathScreen = () => {
             </div>
 
         </div>
+
+    </div>
+    </div>   
+    </ThemeContext.Provider>  
 
     );
 }
