+++ application/src/main/resources/static/js/react-client/src/pages/reward-screen/index.js
@@ -1,10 +1,20 @@
-import React, { useEffect, useState } from "react";
+import React, { createContext, useEffect, useState } from "react";
 
 import Axios from 'axios';
+import ReactSwitch from "react-switch";
 import style from './style.module.scss';
 import { useParams } from "react-router-dom";
 
+export const ThemeContext = createContext(null);
+
 const RewardScreen = () => {
+// light/dark mode
+    const [ theme, setTheme ] = useState("dark");
+    const toggleTheme = () => {
+      setTheme((curr) => (curr === "light" ? "dark" : "light"));
+    };
+// light/dark mode
+    
     // const { id } = useParams();
     const [rewards, setRewards] = useState(null)
     const [loading, setLoading] = useState(true);
@@ -28,6 +38,16 @@ const RewardScreen = () => {
     }, [rewards]);
 
     return (
+// light/dark mode
+        <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      <div className="Contact" id={theme} >   
+      <div className={theme=== 'dark' ? style.dark : style.light}>
+      <div className="switch">
+          {/* <label> {theme === "light" ? "Light Mode" : "Dark Mode"}</label> */}
+        <ReactSwitch onChange={toggleTheme} checked={theme === "dark"}/>
+        </div>
+{/* light/dark mode */}
+
         loading ? <h2>Loading...</h2> :
         <div>
             <section className={style.rewardsSection}>
@@ -45,6 +65,10 @@ const RewardScreen = () => {
         </div>
         </section>
         </div>
+
+    </div>
+    </div>   
+    </ThemeContext.Provider>      
     );
     
 }
