+++ application/src/main/resources/static/js/react-client/src/style.module.scss
@@ -5,16 +5,3 @@ body {
     text-align: center;
 }
 
-.light {
-    background-image: url(../src/assets/background/light-mode-background.png);
-    // background-color: #00e7ff;
-    background-size: cover;
-
-}
-
-.dark {
-    background-image: url(../src/assets/background/dark-mode-background.png);
-    color: white;
-    background-size: cover;
-    // background-color: #101010;
-}
